requirements.txt
PyPDF2==3.0.1
requests==2.32.3
python-dotenv==1.0.1
